split up ofio
